requirements.txt
easyocr
sentence-transformers
faiss-cpu
unstructured[all-docs]
numpy
huggingface_hub[hf_xet]
Ollama

